Revert "HttpCache big boy entrance"
This reverts commit 38f3aca211a310194c0332c983b8343d152f6c69.
