+++ methanol/src/main/java/com/github/mizosoft/methanol/internal/flow/AbstractSubscription.java
@@ -38,9 +38,6 @@ import org.checkerframework.checker.nullness.qual.Nullable;
  *
  * @param <T> item's type
  */
-// TODO log failure when executor throws bec it may be swallowed when rethrown
-//      causing us to start pulling out hair when nothing completes
-@SuppressWarnings("unused") // VarHandle indirection
 public abstract class AbstractSubscription<T> implements Subscription {
 
   /*
