--- methanol/src/test/resources/default_cacheability.csv
@@ -1,46 +1,0 @@
-# See HttpCacheTest.defaultCacheability
-# https://httpwg.org/specs/rfc7231.html#overview.of.status.codes
-# Cacheable by default: 200, 203, 204, 206, 300, 301, 404, 405, 410, 414, and 501
-code, cacheableByDefault
-100, false
-101, false
-200, true
-201, false
-202, false
-203, true
-204, true
-205, false
-# Partial content is not supported
-206, false
-300, true
-301, true
-302, false
-303, false
-304, false
-305, false
-307, false
-400, false
-401, false
-402, false
-403, false
-404, true
-405, true
-406, false
-407, false
-408, false
-409, false
-410, true
-411, false
-412, false
-413, false
-414, true
-415, false
-416, false
-417, false
-426, false
-500, false
-501, true
-502, false
-503, false
-504, false
-505, false
